Initial commit of "mobius_strip" variation originally an Apophysis plugin called "mobius" by slobo777, see http://slobo777.deviantart.com/art/Mobius-Apo-Variation-60525279 for details. Transcribed into Java by chronologicaldot, then further modified by CozyG
Note that the Apophysis "mobius" plugin that "mobius_strip" is derived from is different from the other "mobius" Apophysis plugin that existing JWildfire "mobius" variation is based on.  
"mobius_strip" variation is based on equations for the Mobius strip, whereas "mobius" variation is based on equations for Mobius transformations, which are completely different.