Allow expanding param panel with one param and a ressource
We don't normally want to expand the param panel if there is only one
param since it just lists the param twice. But we do want to expand if
there is also a ressource, since the ressource is listed as a param.
This mostly affects custom_wf_full variations with a single parameter;
the parameter would not be expanded before this change.