Resize the nonlinear params panel when a new variation is added
Fixes a bug where adding a new variation with lots of parameters when
the parameter list was expanded didn't adjust the scrollbar so
additional variations could not be added.

I actually added a call to resizeNonlinearParamsPanel in
refreshParamControls, since resizing the panel (and adjusting the
scrollbar) is part of refreshing the parameter controls.