Changed EPSILON to SMALL_EPSILON when used to avoid division by 0
EPSILON is only 1e-8; SMALL_EPSILON is 1e-300, which is a better value
to avoid division by 0. It also matches Apophysis, where Xform.pas
defines EPS as 1E-300. It won't make any difference for most flames, but
it does for a few that use variations like spherical when values get
close to 0.