Mod to allow dc_crackle_wf and dc_perlin to "paint" patterns
Using dc_linear and dc_hexes_wf as a post variation with pAmount 0 will
paint the variation pattern on a shape with affecting the shape. This
mod allows the same to be done with dc_crackle_wf, dc_cracklep_wf, and
dc_perlin by using the pAffineTP point instead of a random point when
pAmount is 0. This won't affect use as a normal variation since setting
pAmount to 0 in that case results in a single point.

I also changed dc_cracklep_wf to be a subclass of dc_crackle_wf instead
of duplicating the code from crackle. This is not only cleaner, but
allows the mod for dc_crackle_wf and dc_cracklep_wf to be done in one
place.