Support for prepost variations, with prepost_affine and prepost_mobius
Prepost variations are executed both before the other variations (along
with other pre_ variations) and after the other variations (along with
other post_ variations). One of the two will be an inverse function,
calling invtransform instead of transform. Prepost variations are
identified with priority 2 (calls invtransform for pre and transform for
post) or -2 (calls transform for pre and invtransform for post).