Fixed bug with dynamic parameter expansion
After changing a parameter with dynamic parameter expansion on a
variation whose first parameter was a ressource using the parameter
spinner (below the amount spinner, not the expanded list), it would be
disabled. This happened because refreshParamControls was called to
refresh the parameters, which selects the first parameter and disables
the spinner if that is a ressource.

Fixed by adding a parameter to refreshParamControls specifying whether
to select the first parameter so the offending selection can be avoided
in this corner case.