Bug fixes for hdr files as color and displacement maps
Scaling is different for hdr vs rgb files, but was not done correctly
for color and displacement maps. For color maps, hdr pixels are now
multiplied by 255 to scale them the same as rgb pixels. For displacement
maps, rgb pixels are now divided by 255 to scale them the same as hdr
pixels, eliminating the need to scale the displ_amount values in the
variations that use them.