Fix issue with Mirror Translations
Previously, when Mirror Translations was enabled, nothing happened until
a translation was actually changed. But if the flame was saved before
changing a translation, the translations would be mirrored when
reloaded, so the saved version wouldn't match the flame before the save.
This was confusing to users who click to see what happens and just go on
when nothing happens since their flames wouldn't save correctly.

To fix, now when Mirror Translations is enabled it immediately mirrors
the current setting, both fixing the issue and giving users who randomly
click buttons feedback that that button does something.