Add length limit for lsystem_js
The lsystem_js variation can hang the system or cause out of memory
errors if the iters parameter is too large for the grammar. This change
avoids the problem by setting a length limit for the string the lsystem
generates. By default, the limit is 25,000, but users with fast CPUs and
lots of memory can change it in Preferences if needed.