Fix NotSerializableExceptions when saving render state
A number of variations would prevent the Interactive Renderer from being
able to save the render state because they used classes that were not
Serializable, specifically DoubleWrapper (used for sinandcos) and
UVCOlorMapper. Fixed by making UVColorMapper Serializable and creating a
new Serializable class DoubleWrapperWF (since DoubleWrapper is a third
party class) and changing the variations to use it.