Fix bugs in Open Image
1. Vertical placement of image panel in the frame was incorrect, and
unstable while resizing. Fixed by using frame.add instead of
frame.setContentPane.

2. Vertical size of the frame was too small because the menubar height
was omitted.

3. The image didn't fit in the frame after resizing because the new
width didn't account for the frame border.

4. The JTattoo look and feels don't set the frame insets, so the frame
wasn't big enough for the image when using them. Worked around this by
defining alternate insets to use when the real ones are 0. (This isn't
perfect since the insets vary in size; the chosen values seem to work
well for most of the options.)