Subflame fixes
Subflame_wf will hang JWildfire if it tries to open a file that doesn't
exist, whether from pasting a flame from someone else referencing a file
from their computer, or from a non-existent sequence file (usually
because the animation went too many frames).

With this version, when a flame_filename is specified it will
immediately read the file into the flame, just as if it was pasted
there. If flame_is_sequence is not 1, the flame_filename is then
ignored, not even saved; only the flame is used. If flame_is_sequence is
1, then flame_filename is still saved and used to generate the name of
the file to used based on the sequence, but if the file doesn't exist
the flame is used instead. This may result in some unexpected animation
glitches, but JWildfire won't hang.

When loading a flame where flame_filename is specified but that file
doesn't exist, JWildfire will display an error at load time and abort
loading the flame. But if the file exists, it will be loaded as
described, allowing old flames to be loaded.