Coloring changes
User visible:
There are now two coloring types, selectable per transform on the
Color tab:
NONE (the default for final transforms) does not change the color;
this is the way final transforms have traditionally worked in JWildfire

DIFFUSION (the default for normal transforms) uses the Color and
Speed settings with the gradient; this is the standard method for
coloring flames.

Changing the Color Type for final transforms to DIFFUSION allows final
transforms to adjust the colors, as it does in other flame programs.
Changing the Color Type for normal transforms to NONE allows them to
manipulate true color (rgbColor) variations without sucking the color
away.

Internal:
Determining the RGB values for transformed points is now done in the
XForm module using additional TransformationStep classes, which
replace the ColorProvider classes from RenderIterationState. This
sets the stage for adding future coloring types.