Update to coloring changes
Bug fixes and additions to improve compatibility. Biggest change is
addition of UNSET color type used when creating a new xform. It is
changed to the correct default (DIFFUSION for normal and NONE for final)
at start of flame render or when flame editor accesses it. Without this,
the default would not be set correctly in scripts.