Move code to convert RGB to RenderColor so it can be reused
No behavior changes. The code to convert RGB to RenderColor based on
white level was moved from inside the createRenderPalette method to the
RenderColor class in new constructors with a whitelevel parameter. This
allows the code to be reused for other purposes, such as the forthcoming
TARGET coloring type.

As a side note, the pWhiteLevel parameter of createRenderPalette was not
used, so it is ignored in the new RenderColor constructors as well to
ensure no behavior changes. This should be evaluated to determine if it
should actually be used or not.