Add new coloring types: TARGET and TARGETG
Two new coloring types are added:

TARGET ignores the gradient, but uses a specific color that can be set
by clicking a new TargetColor button. The new color is interpolated from
the current color and the target color. Speed works the same as with
DIFFUSION coloring; -1 sets the new color to the target color, 1 doesn't
change the color, and 0 sets the new color halfway between the current
and target colors.

TARGETG works just like TARGET, but the gradient is used to select the
target color, allowing it to be easily changed by changing or shifting
the gradient. But the intermediate gradient colors are not used.

These coloring types can create very smooth color changes in flames, and
also work with variations that set the RGB color values directly.