Add new coloring type: DISTANCE
Uses the gradient indexed by the distance the transform moved the point.
Color determines the starting point and Speed controls how far along the
gradient each unit of distance moves from there. Like other coloring
types, Speed -1 sets the color to the specified color. The input color
is ignored, so unlike the others Speed 1 will still change the color.

DISTANCE is especially effective for coloring flames made with a single
non-blur transform (blur transforms ignore the input point, so the
distance and thus the color of each point is random).