New coloring type: CYCLIC
CYCLIC shifts the gradient index right by the amount specified by Speed,
wrapping as needed to cycle through gradient colors. It doesn't converge
to a color when iterated, so some other transform needs to use DIFFUSION
to get a stable color that CYCLIC can use as a starting point.

CYCLIC is especially useful for coloring tilings.