Bug fixes
Fixed an issue with some controls where switching between a transform or
layer with a motion control enabled and one with it disabled would not
properly enable/disable the control. Specifically the transform weight
control, all Color controls, and WField type specific controls (below
the line), and the layer density and weight controls.

Motion curves for Material speed now work.