Verify OptiX is available before attempting to use it
Although the OptiX denoiser controls are disabled when OptiX is not
available, it is still possible for it to be enabled. For example, by
loading a flame saved on a computer where it is available (such as flame
shared by someone else). Rendering a flame with OptiX enabled but not
available would display a cryptic message that the denoised file isn't
found and not save the rendered image. This fixes that: no message is
displayed and the image is saved (though not denoised).