Add mode parameter to octagon variation
Currently, octagon doesn't match the Apophysis plugin because a brace is
misplaced. But it is useful like that! So I added a mode parameter to
switch between the plugin behavior (mode 0) and the JWildfire behavior
(mode 1), as well as adding several more modes. I also refactored the
math to make it clearer and more efficient.