Bug fix: initialize width and height for new flames
Flames don't work when width or height are 0 since things like computing
pixelsPerUnit divide by them. It isn't normally a problem since all of
the JWildfire code as well as scripts created from flames set the flame
width and height explicitly. But a script that creates a new flame and
neglects to set them won't work; this fixes that.